refactor(lib): 改进一些实现细节
